Replace old build_demo_web.sh with trunk
