+++ .github/workflows/rust.yml
@@ -182,8 +182,8 @@ jobs:
       - uses: actions/checkout@v2
       - uses: EmbarkStudios/cargo-deny-action@v1
 
-  wasm_bindgen:
-    name: wasm-bindgen
+  trunk:
+    name: trunk
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
@@ -193,8 +193,10 @@ jobs:
           toolchain: 1.61.0
           target: wasm32-unknown-unknown
           override: true
-      - run: ./sh/setup_web.sh
-      - run: ./sh/wasm_bindgen_check.sh
+      - name: Download and install Trunk binary
+        run: wget -qO- https://github.com/thedodd/trunk/releases/latest/download/trunk-x86_64-unknown-linux-gnu.tar.gz | tar -xzf-
+      - name: Build
+        run: ./trunk build
 
   android:
     name: android
