+++ assets/CNAME
@@ -1,0 +1,1 @@
+www.egui.rs
