--- docs/CNAME
@@ -1,1 +1,0 @@
-www.egui.rs
