--- docs/README.md
@@ -1,3 +1,0 @@
-This folder contains the files required for the egui web demo hosted at <https://www.egui.rs/>.
-
-The reason the folder is called "docs" is because that is the name that GitHub requires in order to host a web page from the `master` branch of a repository.
