--- docs/example.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-
-<head>
-    <meta http-equiv="refresh" content="0; url = https://www.egui.rs/#http" />
-</head>
-
-<body>
-</body>
-
-</html>
