+++ egui_demo_app/Cargo.toml
@@ -11,9 +11,6 @@ default-run = "egui_demo_app"
 [package.metadata.docs.rs]
 all-features = true
 
-[lib]
-crate-type = ["cdylib", "rlib"]
-
 
 [features]
 default = ["glow", "persistence"]
@@ -70,3 +67,4 @@ tracing-subscriber = "0.3"
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 console_error_panic_hook = "0.1.6"
 tracing-wasm = "0.2"
+web-sys = { version = "0.3", features = [] }
