+++ sh/setup_web.sh
@@ -5,4 +5,4 @@ cd "$script_path/.."
 
 # Pre-requisites:
 rustup target add wasm32-unknown-unknown
-cargo install wasm-bindgen-cli --version 0.2.82
+cargo install --locked trunk
