+++ sh/wasm_size.sh
@@ -3,4 +3,4 @@ set -eu
 script_path=$( cd "$(dirname "${BASH_SOURCE[0]}")" ; pwd -P )
 cd "$script_path"
 
-./build_demo_web.sh && ls -lh ../docs/*.wasm
+trunk build --release && ls -lh ../dist/*.wasm
